monitoring-project/custom_exporter/requirements.txt
prometheus-client==0.17.1
requests==2.31.0
flask==2.3.3
psycopg2-binary==2.9.7  # Add this for PostgreSQL connection